mcps/monday/pyproject.toml
[project]
name = "monday-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.0.2",
    "httpx>=0.28.0",
    "python-dotenv>=1.2.2",
]
